+++ src/main/java/spring/core/entity/discount/DiscountStrategy.java
@@ -3,6 +3,8 @@ package spring.core.entity.discount;
 import spring.core.entity.Event;
 import spring.core.entity.User;
 
+import java.math.BigDecimal;
+
 public interface DiscountStrategy {
-    Double apply(User user, Event event);
+    BigDecimal apply(User user, Event event);
 }
